Move the tuple_map utility into a seaprate header
